url/resolve-urls.xhtml

<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE html>
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>HTML Test: Resolve URLs - xml:base</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" title="2.5.2 Resolving URLs" href="http://www.w3.org/html/wg/drafts/html/CR/infrastructure.html#resolving-urls" />
    <link rel="help" href="http://www.w3.org/TR/xhtml1/" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body xml:base="xml">  
    <div id="log"></div>
    <script>
      <![CDATA[
      //resolve relative URLs
      var a1 = document.createElement("a"),
          base_url = document.location.href,
          host = document.location.host,
          root = base_url.replace(document.location.pathname, "/"),
          path = base_url.substr(0, base_url.lastIndexOf("/"));

      var testData = [
        "/", root,
        "/test.txt", root + "test.txt",
        ".", path + "/",
        "..", path.substr(0, path.lastIndexOf("/") + 1),
        "test.txt", path + "/test.txt",
        "./test.txt", path + "/test.txt",
        "../test.txt", path.substr(0, path.lastIndexOf("/") + 1) + "test.txt",
        "../aaa/test.txt", path.substr(0, path.lastIndexOf("/") + 1) + "aaa/test.txt",
        "\u3167/test.txt", path + "/%E3%85%A7/test.txt"
      ];

      document.body.appendChild(a1);

      for (var i = 0; i < testData.length;i += 2) {
        test(function () {
          a1.setAttribute("href", testData[i]);
          assert_greater_than(a1.href.indexOf(host), -1, "The resolved URL must have a hostname.");
          assert_equals(a1.href, testData[i + 1], "The resolved URL should be " + testData[i + 1] + ".");
        }, "The url '" + testData[i] + "' must be resolved to '" + testData[i + 1] + "'");
      }

      //resolve absolute URLs
      test(function () {
        a1.setAttribute("href", "http://www.example.com");
        assert_equals(a1.href, "http://www.example.com/", "The resolved URL should be 'http://www.example.com/'.");
      }, "The url 'http://www.example.com' must be resolved to 'http://www.example.com/'");
      
      test(function () {
        a1.setAttribute("href", "http://www.ex\u0061mple.com");
        assert_equals(a1.href, "http://www.example.com/", "The resolved URL should be 'http://www.example.com/'.");
      }, "The url 'http://www.ex\\u0061mple.com' must be resolved to 'http://www.example.com/'");

      //Base url is an absolute URL
      document.body.setAttribute("xml:base", "http://www.example.com/test/");
      testData = [
        "/", "http://www.example.com/",
        "/test.txt", "http://www.example.com/test.txt",
        ".", "http://www.example.com/test/",
        "..", "http://www.example.com/",
        "test.txt", "http://www.example.com/test/test.txt",
        "./test.txt", "http://www.example.com/test/test.txt",
        "../test.txt", "http://www.example.com/test.txt",
        "../aaa/test.txt", "http://www.example.com/aaa/test.txt",
        "../../test.txt", "http://www.example.com/test.txt",
        "\u3167/test.txt", "http://www.example.com/test/" + "%E3%85%A7/test.txt",
        "http://www.example2.com/", "http://www.example2.com/"
      ];

      for (var i = 0; i < testData.length; i += 2) {
        test(function () {
          a1.setAttribute("href", testData[i]);
          assert_equals(a1.href, testData[i + 1], "The resolved URL should be " + testData[i + 1] + ".");      
        }, "The url '" + testData[i] + "' must be resolved to '" + testData[i + 1] + "'");
      }
      ]]>
    </script>
  </body>
</html>
